tests/wpt/tests/css/css-anchor-position/anchor-scroll-position-try-014.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#scroll">
<link rel="help" href="https://drafts.csswg.org/css-anchor-1/#fallback-apply">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/test-common.js"></script>
<style>
#cb {
  position: relative;
  width: 200px;
  height: 200px;
  border: solid 1px;
}

#scroller {
  width: 200px;
  height: 200px;
  overflow: scroll;
  display: flex;
  flex-direction: column-reverse;
}

#abspos {
  position: absolute;
  background: hotpink;
  width: 50px;
  height: 50px;

  inset-area: top;
  position-try-fallbacks: flip-block;
  position-anchor: --a;
}

#anchor {
  anchor-name: --a;
  width: 50px;
  min-height: 50px;
  flex: 0;
  margin: 150px 0 150px 50px;
  background: dodgerblue;
}

</style>
<div id="cb">
  <div id="scroller">
    <div id="anchor">
    </div>
  </div>
  <div id="abspos"></div>
</div>

<script>
promise_test(async () => {
  await waitUntilNextAnimationFrame();
  assert_fallback_position(abspos, anchor, 'bottom');
});

promise_test(async () => {
  scroller.scrollTop = -100;
  await waitUntilNextAnimationFrame();
  assert_fallback_position(abspos, anchor, 'top');
});

promise_test(async () => {
  scroller.scrollTop = -50;
  await waitUntilNextAnimationFrame();
  assert_fallback_position(abspos, anchor, 'bottom');
});
</script>
