third_party/blink/web_tests/platform/mac-mac13/virtual/css-anchor-update-enabled/external/wpt/css/css-anchor-position/anchor-scroll-position-try-014-expected.txt

This is a testharness.js-based test.
[FAIL] anchor-scroll-position-try-014 2
  assert_equals: Anchored element should be at the bottom of anchor expected 109 but got 9
Harness: the test ran to completion.
